# Conflicts:
#	app/src/main/java/com/example/your_space/database/DatabaseEntities.kt
#	app/src/main/java/com/example/your_space/network/Constants.kt
#	app/src/main/java/com/example/your_space/network/NetworkServices.kt
#	app/src/main/java/com/example/your_space/network/networkdatamodel/BookingProperty.kt
#	app/src/main/java/com/example/your_space/repository/AppRepository.kt
#	app/src/main/java/com/example/your_space/ui/AppViewModel.kt
#	app/src/main/java/com/example/your_space/ui/ourspaces/SecondFragment.kt

+++ app/src/main/java/com/example/your_space/repository/AppRepository.kt
@@ -1,60 +1,74 @@
 package com.example.your_space.repository
 
+import android.content.Context
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.map
-import com.example.your_space.database.AppDao
-import com.example.your_space.database.bookingToDomainModel
-import com.example.your_space.database.spaceToDomainModel
-import com.example.your_space.network.Network
+import com.example.your_space.database.*
+import com.example.your_space.network.Network.NetworkServices
 import com.example.your_space.network.networkdatamodel.BookingProperty
 import com.example.your_space.network.networkdatamodel.SpaceItemProperty
 import com.example.your_space.network.networkdatamodel.bookingProertyModelToDatabaseModel
 import com.example.your_space.network.networkdatamodel.propertyModelToDatabaseModel
 import com.example.your_space.ui.booking.BookItem
-import com.example.your_space.ui.ourspaces.SpaceItem
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
-class AppRepository(private val database: AppDao) {
+const val REPOSITORY_ERROR_STRING = "Error in repository"
 
-    val workingSpacesRepo: LiveData<List<SpaceItem>> =
-        database.gelAllWorkingSpaces().map { it.spaceToDomainModel() }
+class AppRepository private constructor(private val database: AppDao) {
+
+    val workingSpacesRepo: LiveData<List<WorkingSpaceDB>> =
+        database.gelAllWorkingSpaces()
 
     val BookingsRepo: LiveData<List<BookItem>> =
         database.gelAllBookings().map { it.bookingToDomainModel() }
 
 
     suspend fun refreshAllBookingString(): List<BookingProperty> {
-        return Network.NetworkServices.getAllBookings()
+        return NetworkServices.getAllBookings()
     }
 
     suspend fun refreshAllWorkingSpacesString(): List<SpaceItemProperty> {
-        return Network.NetworkServices.getAllWorkingSpaces()
+        return NetworkServices.getAllWorkingSpaces()
     }
 
     suspend fun refreshWorkingSpaces() {
         try {
-            val workingSpacesList = Network.NetworkServices.getAllWorkingSpaces()
+            val workingSpacesList = NetworkServices.getAllWorkingSpaces()
             if (workingSpacesList.isNotEmpty()) {
                 withContext(Dispatchers.IO) {
                     database.insertAllWorkingSpaces(*(workingSpacesList.propertyModelToDatabaseModel()))
                 }
             }
         } catch (e: Exception) {
+            Log.e(REPOSITORY_ERROR_STRING, e.stackTraceToString())
+        }
+    }
 
+    suspend fun loadWorkingSpacesOfPage(pageNumber: Int) {
+        try {
+            val workingSpacesList = NetworkServices.getWorkingSpacesUsingPaging(pageNumber)
+            if (workingSpacesList.isNotEmpty()) {
+                withContext(Dispatchers.IO) {
+                    database.insertAllWorkingSpaces(*(workingSpacesList.propertyModelToDatabaseModel()))
+                }
+            }
+        } catch (e: Exception) {
+            Log.e(REPOSITORY_ERROR_STRING, e.stackTraceToString())
         }
     }
 
     suspend fun refreshBookings() {
         try {
-            val bookingsList = Network.NetworkServices.getAllBookings()
+            val bookingsList = NetworkServices.getAllBookings()
             if (bookingsList.isNotEmpty()) {
                 withContext(Dispatchers.IO) {
                     database.insertAllBookings(*(bookingsList.bookingProertyModelToDatabaseModel()))
                 }
             }
         } catch (e: Exception) {
-
+            Log.e(REPOSITORY_ERROR_STRING, e.stackTraceToString())
         }
     }
 
@@ -65,9 +79,21 @@ class AppRepository(private val database: AppDao) {
     }
 
     suspend fun deleteBookingWithId(bookItem: BookItem) {
-        withContext(Dispatchers.IO) {
-            database.deleteBooking(bookItem.bookId)
+        try {
+            // Do the DELETE request and get response
+            val response = NetworkServices.cancelBooking(bookItem.bookId.toInt())
+            withContext(Dispatchers.IO) {
+                if (response.isSuccessful) {
+                    database.deleteBooking(bookItem.bookId)
+
+                } else {
+                    Log.e("RETROFIT_ERROR", response.code().toString())
+                }
+            }
+        } catch (e: Exception) {
+            Log.e(REPOSITORY_ERROR_STRING, e.stackTraceToString())
         }
+
     }
 
     suspend fun deleteAllBookings() {
@@ -76,4 +102,22 @@ class AppRepository(private val database: AppDao) {
         }
     }
 
+    companion object {
+        @Volatile
+        private var repositoryINSTANCE: AppRepository? = null
+
+        fun getInstance(context: Context): AppRepository {
+            synchronized(this) {
+                var localInstance = repositoryINSTANCE
+
+                if (localInstance == null) {
+                    val appDao = AppDatabase.getInstance(context).dao
+                    localInstance = AppRepository(appDao)
+
+                    repositoryINSTANCE = localInstance
+                }
+                return localInstance
+            }
+        }
+    }
 }
